Make SOLID
