--- Exp_01_Non_SOLID/src/PaymentServices/OnlineOrderService.java
@@ -1,13 +1,0 @@
-package PaymentServices;
-
-public class OnlineOrderService implements OrderService {
-    @Override
-    public void orderRegister(String customerName) {
-        System.out.println("online order registered for " + customerName);
-    }
-
-    @Override
-    public void orderPayment(int foodPrice) {
-        System.out.println("online Payment with Price : " + foodPrice + " Tomans!");
-    }
-}
