--- Exp_01_Non_SOLID/src/PaymentServices/OrderService.java
@@ -1,6 +1,0 @@
-package PaymentServices;
-
-public interface OrderService {
-    void orderRegister(String customerName);
-    void orderPayment(int foodPrice);
-}
