--- Exp_01_Non_SOLID/src/PaymentServices/PhoneOrderService.java
@@ -1,15 +1,0 @@
-package PaymentServices;
-
-
-public class PhoneOrderService implements OrderService{
-    @Override
-    public void orderRegister(String customerName) {
-        System.out.println("Phone order registered for " + customerName);
-    }
-
-    @Override
-    public void orderPayment(int foodPrice) {
-        System.out.println("Phone Payment with Price : " + foodPrice + " Tomans!");
-    }
-
-}
