+++ Step_1_Non_SOLID/src/PaymentServices/OnlineOrderService.java
@@ -1,0 +1,35 @@
+package PaymentServices;
+
+public class OnlineOrderService implements OrderService {
+    @Override
+    public void onSiteOrderRegister(String customerName) {
+        //Empty Body!
+    }
+
+    @Override
+    public void onlineOrderRegister(String customerName) {
+        System.out.println("online order registered for " + customerName);
+    }
+
+    @Override
+    public void onSiteOrderPayment(int foodPrice) {
+        //Empty Body
+    }
+
+    @Override
+    public void onlineOrderPayment(int foodPrice) {
+        System.out.println("online Payment with Price : " + foodPrice + " Tomans!");
+    }
+
+    @Override
+    public void onPhoneOrderRegister(String customerName) {
+        //Empty Body!
+    }
+
+    @Override
+    public void onPhoneOrderPayment(int foodPrice) {
+        //Empty Body!
+    }
+    
+
+}
