+++ Step_3_With_SOLID/src/PaymentServices/OnlineOrderService.java
@@ -2,34 +2,12 @@ package PaymentServices;
 
 public class OnlineOrderService implements OrderService {
     @Override
-    public void onSiteOrderRegister(String customerName) {
-        //Empty Body!
-    }
-
-    @Override
-    public void onlineOrderRegister(String customerName) {
+    public void orderRegister(String customerName) {
         System.out.println("online order registered for " + customerName);
     }
 
     @Override
-    public void onSiteOrderPayment(int foodPrice) {
-        //Empty Body
-    }
-
-    @Override
-    public void onlineOrderPayment(int foodPrice) {
+    public void orderPayment(int foodPrice) {
         System.out.println("online Payment with Price : " + foodPrice + " Tomans!");
     }
-
-    @Override
-    public void onPhoneOrderRegister(String customerName) {
-        //Empty Body!
-    }
-
-    @Override
-    public void onPhoneOrderPayment(int foodPrice) {
-        //Empty Body!
-    }
-    
-
 }
