+++ Step_3_With_SOLID/src/PaymentServices/OrderService.java
@@ -1,11 +1,6 @@
 package PaymentServices;
 
 public interface OrderService {
-    public void onSiteOrderRegister(String customerName);
-    public void onlineOrderRegister(String customerName);
-    public void onPhoneOrderRegister(String customerName);
-    public void onSiteOrderPayment(int foodPrice);
-    public void onlineOrderPayment(int foodPrice);
-    public void onPhoneOrderPayment(int foodPrice);
-
+    void orderRegister(String customerName);
+    void orderPayment(int foodPrice);
 }
