+++ Step_3_With_SOLID/src/PaymentServices/PhoneOrderService.java
@@ -3,32 +3,12 @@ package PaymentServices;
 
 public class PhoneOrderService implements OrderService{
     @Override
-    public void onSiteOrderRegister(String customerName) {
-        //Empty Body!
-    }
-
-    @Override
-    public void onlineOrderRegister(String customerName) {
-        //Empty Body
-    }
-
-    @Override
-    public void onSiteOrderPayment(int foodPrice) {
-        //Empty Body
-    }
-
-    @Override
-    public void onlineOrderPayment(int foodPrice) {
-        //Empty Body!
-    }
-
-    @Override
-    public void onPhoneOrderRegister(String customerName) {
+    public void orderRegister(String customerName) {
         System.out.println("Phone order registered for " + customerName);
     }
 
     @Override
-    public void onPhoneOrderPayment(int foodPrice) {
+    public void orderPayment(int foodPrice) {
         System.out.println("Phone Payment with Price : " + foodPrice + " Tomans!");
     }
